requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
bidict==0.23.1
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
chainlit==1.3.1
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.14
fastapi==0.112.4
filetype==1.2.0
frozenlist==1.5.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-aws==0.2.4
langchain-core==0.3.13
langchain-text-splitters==0.3.0
langsmith==0.1.137
Lazify==0.4.0
literalai==0.0.623
marshmallow==3.23.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.10
packaging==23.2
propcache==0.2.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.3
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.37.2
syncer==2.0.3
tenacity==9.0.0
tomli==2.0.2
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.27.0
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.16.0
zipp==3.20.2
peewee==3.14.4
langchain
langchain_community
langchain_openai
libmagic
python-magic-bin
chromadb
pypdf